feat(http): move back to rpc
