+++ net/http/rpc/error.go
@@ -1,4 +1,4 @@
-package http
+package rpc
 
 import (
 	"net/http"
