+++ test/rpc.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/alexfalkowski/go-service/meta"
-	nh "github.com/alexfalkowski/go-service/net/http"
+	"github.com/alexfalkowski/go-service/net/http/rpc"
 	v1 "github.com/alexfalkowski/go-service/test/greet/v1"
 )
 
@@ -19,7 +19,7 @@ type Response struct {
 
 type SuccessHandler struct{}
 
-func (*SuccessHandler) Handle(ctx nh.Context, r *Request) (*Response, error) {
+func (*SuccessHandler) Handle(ctx rpc.Context, r *Request) (*Response, error) {
 	name := ctx.Request().URL.Query().Get("name")
 	if name == "" {
 		name = r.Name
@@ -32,16 +32,16 @@ func (*SuccessHandler) Handle(ctx nh.Context, r *Request) (*Response, error) {
 
 type ProtobufHandler struct{}
 
-func (*ProtobufHandler) Handle(_ nh.Context, r *v1.SayHelloRequest) (*v1.SayHelloResponse, error) {
+func (*ProtobufHandler) Handle(_ rpc.Context, r *v1.SayHelloRequest) (*v1.SayHelloResponse, error) {
 	return &v1.SayHelloResponse{Message: "Hello " + r.GetName()}, nil
 }
 
-func (*ProtobufHandler) Error(_ nh.Context, err error) *v1.SayHelloResponse {
+func (*ProtobufHandler) Error(_ rpc.Context, err error) *v1.SayHelloResponse {
 	return &v1.SayHelloResponse{Message: err.Error()}
 }
 
 type ErrorHandler struct{}
 
-func (*ErrorHandler) Handle(_ nh.Context, _ *Request) (*Response, error) {
-	return nil, nh.Error(http.StatusServiceUnavailable, "ohh no")
+func (*ErrorHandler) Handle(_ rpc.Context, _ *Request) (*Response, error) {
+	return nil, rpc.Error(http.StatusServiceUnavailable, "ohh no")
 }
